Simplify log_failed_logins implementation
While the redirect using 302 after login is a reasonable convention we
don't need to depend on it and instead can use the request.user instance
which auth.login will set for us.

Oh, and added mock as a dependency because syslogging in tests is ugly.
Maybe switch to python logging? :)