Fix test.sh for >=Django-2.2
patterns() have been replaced with simple lists but the root urlconf
has to be hashable so we use a tuple.
The patterns are expected to be instances of path/re_path.
print is a function in python3.
osso.relation has been removed.
The test_settings bytecode is written to __pycache__ in python3.